Fix incorrect include path for pgxp.
It only happened to work in some build environments because some of them set up an include dir pointing to /deps/ -- in which case /deps/../pgxp happens to point to the "right place."  The correct style is to reference either relative to the including file, or relative to a well-defined include dir location. Preferred is relative to current file.