hle: preliminary DeliverEvent implementation, with explanation of stack-based state management.
PCSX relied on re-entrant interpreter behavior to invoke events, which is problematic and dangerous. They worked around some dangers by hacking it to only allow one degree of reentrant operation (it would crash on a second level of nesting). For most games this was probably OK. In any case, Mednafen's interpreter is not friendly to re-entrant behavior, and would need significant modification to allow such a thing.